fix(observability): address review feedback
