+++ bin/secret-service/Cargo.toml
@@ -9,8 +9,10 @@ workspace = true
 [dependencies]
 secret-service-proto.workspace = true
 secret-service-server.workspace = true
+strata-bridge-common.workspace = true
 strata-bridge-key-deriv.workspace = true
 strata-bridge-primitives.workspace = true
+strata-metrics.workspace = true
 
 make_buf = { git = "https://github.com/alpenlabs/make_buf", version = "1.0.0" }
 
@@ -19,6 +21,7 @@ cache-advisor = "1.0.16"
 colored = "3.1.1"
 hkdf.workspace = true
 libp2p-identity = { workspace = true, features = ["ed25519"] }
+metrics.workspace = true
 musig2.workspace = true
 rand.workspace = true
 rcgen = "0.13.2"
@@ -29,7 +32,6 @@ terrors.workspace = true
 tokio.workspace = true
 toml = "0.8.19"
 tracing.workspace = true
-tracing-subscriber.workspace = true
 
 [target.'cfg(not(target_env = "msvc"))'.dependencies]
 memory_pprof = { path = "../../crates/memory_pprof", optional = true }
