+++ bin/secret-service/src/config.rs
@@ -19,6 +19,22 @@ pub struct Config {
 
     /// Which bitcoin network to use
     pub network: Option<Network>,
+
+    /// Configuration for process-level metrics exporters.
+    #[serde(default)]
+    pub metrics: MetricsConfig,
+}
+
+/// Configuration for secret-service process metrics exporters.
+#[derive(Debug, Clone, Default, Eq, PartialEq, Deserialize)]
+pub struct MetricsConfig {
+    /// Optional OTLP endpoint URL for metrics export.
+    ///
+    /// If unset, metrics reuse `STRATA_BRIDGE_OTLP_URL` when present.
+    pub otlp_url: Option<String>,
+
+    /// Optional Prometheus listener address, for example `0.0.0.0:9616`.
+    pub prometheus_listener_addr: Option<SocketAddr>,
 }
 
 /// Configuration for the transport layer.
