+++ bin/secret-service/src/main.rs
@@ -2,6 +2,7 @@
 
 // use secret_service_server::rustls::ServerConfig;
 pub mod config;
+mod observability;
 pub mod seeded_impl;
 #[cfg(test)]
 mod tests;
@@ -17,7 +18,8 @@ use seeded_impl::Service;
 #[cfg(not(target_env = "msvc"))]
 use tikv_jemallocator::Jemalloc;
 use tls::load_tls;
-use tracing::{info, warn, Level};
+use tokio::runtime::Handle;
+use tracing::{info, warn};
 
 #[cfg(not(target_env = "msvc"))]
 #[global_allocator]
@@ -42,17 +44,29 @@ pub static DEV_MODE: LazyLock<bool> =
 async fn main() {
     #[cfg(feature = "memory_profiling")]
     memory_pprof::setup_memory_profiling(3_000);
-    tracing_subscriber::fmt().with_max_level(Level::INFO).init();
-    tls::install_rustls_crypto_provider();
-    if *DEV_MODE {
-        warn!("⚠️ DEV_MODE active");
-    }
     let config_path =
         PathBuf::from_str(&args().nth(1).unwrap_or_else(|| "config.toml".to_string()))
             .expect("valid config path");
 
     let text = std::fs::read_to_string(&config_path).expect("read config file");
     let conf: Config = toml::from_str(&text).expect("valid toml");
+    let network = conf.network.unwrap_or(Network::Signet);
+    let seed_path = conf
+        .seed
+        .unwrap_or(PathBuf::from_str("seed").expect("valid path"));
+
+    tls::install_rustls_crypto_provider();
+    observability::init(
+        &conf.metrics,
+        &conf.tls,
+        network,
+        *DEV_MODE,
+        &Handle::current(),
+    );
+    if *DEV_MODE {
+        warn!("DEV_MODE active");
+    }
+
     let tls = load_tls(conf.tls).await;
 
     let config = ServerConfig {
@@ -61,19 +75,16 @@ async fn main() {
         connection_limit: conf.transport.conn_limit,
     };
 
-    let service = Service::load_from_seed(
-        &conf
-            .seed
-            .unwrap_or(PathBuf::from_str("seed").expect("valid path")),
-        conf.network.unwrap_or(Network::Signet),
-    )
-    .await
-    .expect("good service");
+    let service = Service::load_from_seed(&seed_path, network)
+        .await
+        .expect("good service");
 
     info!("Running on {}", config.addr.to_string().bold());
     match config.connection_limit {
         Some(conn_limit) => info!("Connection limit: {}", conn_limit.to_string().bold()),
         None => info!("No connection limit"),
     }
-    run_server(config, service.into()).await.unwrap();
+    let result = run_server(config, service.into()).await;
+    observability::finalize();
+    result.unwrap();
 }
